blake2.cpp -> blake2.c
